Move SAT allocation into rule helpers
